:recycle: 스코프 축소,storeManagerController -> storeController 통합
