--- src/main/java/com/potato/ecommerce/domain/store/dto/OrderAcceptResponse.java
@@ -1,12 +1,0 @@
-package com.potato.ecommerce.domain.store.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class OrderAcceptResponse {
-
-    private String message;
-    private String description;
-}
